Revert "Move license server code from shared to backend (#3848)"
This reverts commit 0897728714e8f00d1f19ab97121f9f6ed51abc29.
