--- packages/back-end/src/enterprise/index.ts
@@ -1,1 +1,0 @@
-export * from "./licenseUtil";
