+++ packages/back-end/src/queryRunners/ExperimentResultsQueryRunner.ts
@@ -1,4 +1,7 @@
-import { analyzeExperimentPower } from "shared/enterprise";
+import {
+  analyzeExperimentPower,
+  orgHasPremiumFeature,
+} from "shared/enterprise";
 import { addDays } from "date-fns";
 import {
   expandMetricGroups,
@@ -11,7 +14,6 @@ import {
 import { FALLBACK_EXPERIMENT_MAX_LENGTH_DAYS } from "shared/constants";
 import { daysBetween } from "shared/dates";
 import chunk from "lodash/chunk";
-import { orgHasPremiumFeature } from "back-end/src/enterprise";
 import { ApiReqContext } from "back-end/types/api";
 import {
   ExperimentSnapshotAnalysis,
