+++ packages/back-end/src/routers/teams/teams.controller.ts
@@ -1,6 +1,6 @@
 import type { Response } from "express";
 import { areProjectRolesValid, isRoleValid } from "shared/permissions";
-import { orgHasPremiumFeature } from "back-end/src/enterprise";
+import { orgHasPremiumFeature } from "shared/enterprise";
 import { TeamInterface } from "back-end/types/team";
 import {
   createTeam,
