+++ packages/back-end/src/services/experiments.ts
@@ -37,9 +37,9 @@ import {
   isFactMetricId,
   isMetricJoinable,
 } from "shared/experiments";
+import { orgHasPremiumFeature } from "shared/enterprise";
 import { hoursBetween } from "shared/dates";
 import { v4 as uuidv4 } from "uuid";
-import { orgHasPremiumFeature } from "back-end/src/enterprise";
 import { MetricPriorSettings } from "back-end/types/fact-table";
 import {
   ExperimentAnalysisSummaryVariationStatus,
