+++ packages/back-end/src/services/organizations.ts
@@ -3,6 +3,14 @@ import { freeEmailDomains } from "free-email-domains-typescript";
 import { cloneDeep } from "lodash";
 import { Request } from "express";
 import {
+  getLicense,
+  isActiveSubscriptionStatus,
+  isAirGappedLicenseKey,
+  licenseInit,
+  postSubscriptionUpdateToLicenseServer,
+  getSubscriptionFromLicense,
+} from "shared/enterprise";
+import {
   areProjectRolesValid,
   isRoleValid,
   getDefaultRole,
@@ -79,14 +87,6 @@ import {
   getLicenseMetaData,
   getUserCodesForOrg,
 } from "back-end/src/services/licenseData";
-import {
-  isAirGappedLicenseKey,
-  getLicense,
-  isActiveSubscriptionStatus,
-  getSubscriptionFromLicense,
-  postSubscriptionUpdateToLicenseServer,
-  licenseInit,
-} from "back-end/src/enterprise";
 import {
   encryptParams,
   getSourceIntegrationObject,
