+++ packages/shared/src/enterprise/public-key.ts
