+++ packages/shared/test/enterprise.test.ts
@@ -1,4 +1,4 @@
-import { getAccountPlan } from "back-end/src/enterprise";
+import { getAccountPlan } from "../src/enterprise";
 
 describe("Entitlements", () => {
   it("Determines account plan", () => {
