+++ packages/shared/test/license.test.ts
@@ -1,7 +1,6 @@
 import fetch, { Response } from "node-fetch";
 
 import cloneDeep from "lodash/cloneDeep";
-import { LicenseInterface } from "shared/enterprise";
 import {
   LICENSE_SERVER_URL,
   licenseInit,
@@ -9,8 +8,9 @@ import {
   resetInMemoryLicenseCache,
   getLicenseError,
   backgroundUpdateLicenseFromServerForTests,
-} from "back-end/src/enterprise/licenseUtil";
-import * as LicenseModelModule from "back-end/src/enterprise/models/licenseModel";
+  LicenseInterface,
+} from "../src/enterprise/licenseUtil";
+import * as LicenseModelModule from "../src/enterprise/models/licenseModel";
 
 const LicenseModel = LicenseModelModule.LicenseModel;
